+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Compatibility:
 * Declare `File::SHARE_DELETE` constant (#3745, @andrykonchin).
 * Support `symbolize_names` argument to `MatchData#named_captures` (#3681, @rwstauner).
 * Support `Proc#initialize_{dup,copy}` for subclasses (#3681, @rwstauner).
+* Remove deprecated `Encoding#replicate` method (#3681, @rwstauner).
 
 Performance: